Make discovery v5 available in full mode
